refactor : inputValidator 기능 위임
