+++ src/main/java/view/InputView.java
@@ -1,23 +1,21 @@
 package view;
 
 import dto.CarNameRequest;
+import dto.CountRequest;
 import java.util.Arrays;
 import java.util.Scanner;
 
 public class InputView {
     Scanner scanner = new Scanner(System.in);
-    InputValidator inputValidator = new InputValidator();
-
     public CarNameRequest readCars() {
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
         String input = scanner.nextLine().trim();
         return new CarNameRequest(input);
     }
 
-    public int readCount() {
+    public CountRequest readCount() {
         System.out.println("시도할 회수는 몇회인가요?");
         String input = scanner.nextLine().trim();
-        inputValidator.validateCount(input);
-        return Integer.parseInt(input);
+        return new CountRequest(input);
     }
 }
